up 样式

+++ src/app/routes/sys-message/message-list/message-list.component.ts
@@ -1,20 +1,21 @@
 import { Component, OnInit } from '@angular/core';
-import {MessageInformService} from "../message-inform.service";
-import {PageEvent} from "../../../shared/directives/ng2-datatable/DataTable";
-import {Page} from "../../../core/page/page";
-import {HeaderComponent} from "../../../layout/header/header.component";
 import {Router} from "@angular/router";
 import {isNullOrUndefined} from "util";
+import {HeaderComponent} from "../../../layout/header/header.component";
+import {PageEvent} from "../../../shared/directives/ng2-datatable/DataTable";
+import {Page} from "../../../core/page/page";
+import {AppComponent} from "../../../app.component";
+import {MessageInformService} from "../sys-message.service";
 
 const swal = require('sweetalert');
 declare var $: any;
 
 @Component({
-  selector: 'app-message-inform',
-  templateUrl: './message-inform.component.html',
-  styleUrls: ['./message-inform.component.scss']
+  selector: 'app-message-list',
+  templateUrl: './message-list.component.html',
+  styleUrls: ['./message-list.component.scss']
 })
-export class MessageInformComponent implements OnInit {
+export class MessageListComponent implements OnInit {
   public orderType: string = '';
   private platformInfoData:any;                           //平台消息的数据
   private deletebutton:Object;                            //删除按钮
@@ -73,20 +74,24 @@ export class MessageInformComponent implements OnInit {
    *1.首先获取到当前选择的id
    */
   updateMoreIsRead(curPage){
-    let url='/notifyAgent/updateIsRead';
     let obj=$("._every[checked='checked']");
-    for(let i=0;i<obj.length;i++){
-      this.idArr.push($(obj[i]).val())
-    }
-    let idStr= this.idArr.join(',');
-    let data={
-      ids	:idStr
-    };
-    this.platformInfoData=this.messageInformService.updateproblem(url,data);
-    this.queryAdminNotify(curPage);
-    if($('._all').prop("checked")){//如果全选按钮被勾选了，让它没有勾选
-      $('._all').prop("checked",false);
-      $('._all').attr("checked",false);
+    if(obj.length==0){
+      AppComponent.rzhAlt("info", '请先进行选择');
+    }else{
+      let url='/notifyAgent/updateIsRead';
+      for(let i=0;i<obj.length;i++){
+        this.idArr.push($(obj[i]).val())
+      }
+      let idStr= this.idArr.join(',');
+      let data={
+        ids	:idStr
+      };
+      this.platformInfoData=this.messageInformService.updateproblem(url,data);
+      this.queryAdminNotify(curPage);
+      if($('._all').prop("checked")){//如果全选按钮被勾选了，让它没有勾选
+        $('._all').prop("checked",false);
+        $('._all').attr("checked",false);
+      }
     }
   }
 
@@ -170,33 +175,37 @@ export class MessageInformComponent implements OnInit {
    */
   delMoreInfo(curPage){
     let that=this;
-    swal({
-      title: "您确定要删除吗？",
-      text: "您确定要删除这条消息？",
-      type: "warning",
-      showCancelButton: true,
-      cancelButtonText: '取消',
-      closeOnConfirm: false,
-      confirmButtonText: "确认",
-      confirmButtonColor: "#ec6c62"
-    },function(isConfirm){
-      if (isConfirm) {
-        swal.close(); //关闭弹框
-        let url='/notifyAgent/deleteByIds';
-        let obj=$("._every[checked='checked']");
-        for(let i=0;i<obj.length;i++){
-          that.idArr.push($(obj[i]).val())
+    if($("._every[checked='checked']").length==0){
+      AppComponent.rzhAlt("info", '请先进行选择');
+    }else{
+      swal({
+        title: "您确定要删除吗？",
+        text: "您确定要删除这条消息？",
+        type: "warning",
+        showCancelButton: true,
+        cancelButtonText: '取消',
+        closeOnConfirm: false,
+        confirmButtonText: "确认",
+        confirmButtonColor: "#ec6c62"
+      },function(isConfirm){
+        if (isConfirm) {
+          swal.close(); //关闭弹框
+          let url='/notifyAgent/deleteByIds';
+          let obj=$("._every[checked='checked']");
+          for(let i=0;i<obj.length;i++){
+            that.idArr.push($(obj[i]).val())
+          }
+          let idStr= that.idArr.join(',');
+          let data={
+            ids:idStr
+          }
+          that.messageInformService.delRequest(url,data)
+          that.queryAdminNotify(curPage);
+          that.headerComponent.queryNotify();
+          $("._all").prop('checked',false);
+          $("._all").attr('checked',false);
         }
-        let idStr= that.idArr.join(',');
-        let data={
-          ids:idStr
-        }
-        that.messageInformService.delRequest(url,data)
-        that.queryAdminNotify(curPage);
-        that.headerComponent.queryNotify();
-        $("._all").prop('checked',false);
-        $("._all").attr('checked',false);
-      }
-    });
+      });
+    }
   }
 }
